x/superchain/types/messages_python_code_test.go
package types

import (
	"testing"

	"github.com/Peixer/superchain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreatePythonCode_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreatePythonCode
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreatePythonCode{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreatePythonCode{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdatePythonCode_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdatePythonCode
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdatePythonCode{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdatePythonCode{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeletePythonCode_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeletePythonCode
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeletePythonCode{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeletePythonCode{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
